requirements.txt

ollama
colorama
chromadb
pygments
duckduckgo_search
pywin32
pyperclip
beautifulsoup4
requests
pyusb
